+++ 1.18/Common/src/main/java/fuzs/illagerinvasion/core/CommonAbstractions.java
@@ -1,5 +1,6 @@
 package fuzs.illagerinvasion.core;
 
+import fuzs.illagerinvasion.helper.IllagerSpellEnumHelper;
 import fuzs.puzzleslib.api.core.v1.ServiceProviderHelper;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.raid.Raider;
@@ -9,5 +10,7 @@ public interface CommonAbstractions {
 
     Object registerRaiderType(String internalName, EntityType<? extends Raider> entityType, int[] spawnsPerWaveBeforeBonus);
 
-    Object registerIllagerSpell(String internalName, double spellColorRed, double spellColorGreen, double spellColorBlue);
+    default Object registerIllagerSpell(String internalName, double spellColorRed, double spellColorGreen, double spellColorBlue) {
+        return IllagerSpellEnumHelper.registerIllagerSpell(internalName, spellColorRed, spellColorGreen, spellColorBlue);
+    }
 }