+++ 1.18/Forge/src/main/java/fuzs/illagerinvasion/data/ModEntityTypeTagProvider.java
@@ -14,5 +14,6 @@ public class ModEntityTypeTagProvider extends AbstractTagProvider.EntityTypes {
     @Override
     protected void addTags() {
         this.tag(EntityTypeTags.RAIDERS).add(ModRegistry.BASHER_ENTITY_TYPE.get(), ModRegistry.PROVOKER_ENTITY_TYPE.get(), ModRegistry.NECROMANCER_ENTITY_TYPE.get(), ModRegistry.SORCERER_ENTITY_TYPE.get(), ModRegistry.ARCHIVIST_ENTITY_TYPE.get(), ModRegistry.MARAUDER_ENTITY_TYPE.get(), ModRegistry.INQUISITOR_ENTITY_TYPE.get(), ModRegistry.ALCHEMIST_ENTITY_TYPE.get());
+        this.tag("numismatic-overhaul:the_bourgeoisie").add(ModRegistry.BASHER_ENTITY_TYPE.get(), ModRegistry.PROVOKER_ENTITY_TYPE.get(), ModRegistry.NECROMANCER_ENTITY_TYPE.get(), ModRegistry.SORCERER_ENTITY_TYPE.get(), ModRegistry.ARCHIVIST_ENTITY_TYPE.get(), ModRegistry.MARAUDER_ENTITY_TYPE.get(), ModRegistry.INQUISITOR_ENTITY_TYPE.get(), ModRegistry.ALCHEMIST_ENTITY_TYPE.get());
     }
 }