+++ 1.20/Common/src/main/java/fuzs/illagerinvasion/client/IllagerInvasionClient.java
@@ -16,6 +16,7 @@ import net.minecraft.client.model.geom.LayerDefinitions;
 import net.minecraft.client.model.geom.builders.CubeDeformation;
 import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.client.particle.FlameParticle;
+import net.minecraft.client.particle.HeartParticle;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
@@ -60,6 +61,7 @@ public class IllagerInvasionClient implements ClientModConstructor {
     @Override
     public void onRegisterParticleProviders(ParticleProvidersContext context) {
         context.registerParticleFactory(ModRegistry.MAGIC_FLAME_PARTICLE_TYPE.get(), FlameParticle.Provider::new);
+        context.registerParticleFactory(ModRegistry.NECROMANCER_BUFF_PARTICLE_TYPE.get(), HeartParticle.Provider::new);
     }
 
     @Override