+++ 1.20/Forge/src/main/java/fuzs/illagerinvasion/data/ModEntityTypeTagProvider.java
@@ -16,5 +16,6 @@ public class ModEntityTypeTagProvider extends AbstractTagProvider.EntityTypes {
     protected void addTags(HolderLookup.Provider provider) {
         this.tag(EntityTypeTags.RAIDERS).add(ModRegistry.BASHER_ENTITY_TYPE.get(), ModRegistry.PROVOKER_ENTITY_TYPE.get(), ModRegistry.NECROMANCER_ENTITY_TYPE.get(), ModRegistry.SORCERER_ENTITY_TYPE.get(), ModRegistry.ARCHIVIST_ENTITY_TYPE.get(), ModRegistry.MARAUDER_ENTITY_TYPE.get(), ModRegistry.INQUISITOR_ENTITY_TYPE.get(), ModRegistry.ALCHEMIST_ENTITY_TYPE.get());
         this.tag(EntityTypeTags.FALL_DAMAGE_IMMUNE).add(ModRegistry.INVOKER_ENTITY_TYPE.get());
+        this.tag("numismatic-overhaul:the_bourgeoisie").add(ModRegistry.BASHER_ENTITY_TYPE.get(), ModRegistry.PROVOKER_ENTITY_TYPE.get(), ModRegistry.NECROMANCER_ENTITY_TYPE.get(), ModRegistry.SORCERER_ENTITY_TYPE.get(), ModRegistry.ARCHIVIST_ENTITY_TYPE.get(), ModRegistry.MARAUDER_ENTITY_TYPE.get(), ModRegistry.INQUISITOR_ENTITY_TYPE.get(), ModRegistry.ALCHEMIST_ENTITY_TYPE.get());
     }
 }