+++ 1.20/CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v8.0.1-1.20.1] - 2023-08-11
+### Changed
+- Added illagers to the `numismatic-overhaul:the_bourgeoisie` entity type tag
+### Fixed
+- Fixed necromancer summoning particles not rendering
+
 ## [v8.0.0-1.20.1] - 2023-07-20
 - Initial release